Rename batch executor and update docs
